refactor(ProductCard): mock data 파일을 직접 의존하지 않도록 타입 분리
