+++ src/types/TrendingGiftsType.ts
@@ -1,0 +1,19 @@
+export type TrendingGiftsType = {
+  id: number;
+  name: string;
+  imageURL: string;
+  price: PriceInfoType;
+  brandInfo: BrandInfoType;
+};
+
+type PriceInfoType = {
+  basicPrice: number;
+  discountRate: number;
+  sellingPrice: number;
+};
+
+type BrandInfoType = {
+  id: number;
+  name: string;
+  imageURL: string;
+};
